% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_distances_functions.R
\docType{class}
\name{distref.data-class}
\alias{distref.data-class}
\title{Class for distref.data}
\description{
Class for distref.data
}
\section{Slots}{

\describe{
\item{\code{data.pt.N}}{numeric.}

\item{\code{NA.pos}}{numeric}

\item{\code{NA.dist}}{numeric}
}}

