% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_distances_functions.R
\name{distref.data}
\alias{distref.data}
\title{Get characteristics of data regarding the reference raster}
\usage{
distref.data(data.pt, r.ref, closest = FALSE,
  longlat = raster::isLonLat(r.ref))
}
\arguments{
\item{data.pt}{Point data set. May be a 2-col matrix of (x,y) coordinates or a
SpatialPointDataFrame. Should be in same projection than r.ref}

\item{r.ref}{Reference raster}

\item{closest}{if a point is over an obstacle, set to TRUE to find the closest
non-obstacle r.ref cell}

\item{longlat}{Logical. if FALSE, Euclidean distance, if TRUE Great Circle
distance. Defined from r.ref.}
}
\description{
Get characteristics of data regarding the reference raster
}
