% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_distances_functions.R
\docType{class}
\name{distref.raster-class}
\alias{distref.raster-class}
\title{Class for distref.raster}
\description{
Class for distref.raster
}
\section{Slots}{

\describe{
\item{\code{path.w}}{}
}}

\examples{

}
