% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_distances_functions.R
\name{Find.ngb.wo.obstacle}
\alias{Find.ngb.wo.obstacle}
\title{Find neighbours without obstacles in different directions}
\usage{
Find.ngb.wo.obstacle(i, r.ref, coords.r, r.ref.NA.nb, cell.Relpos.ngb,
  cell.Relpos.cross)
}
\arguments{
\item{i}{cell number}

\item{r.ref}{reference raster}

\item{coords.r}{coordinates of the reference raster}

\item{r.ref.NA.nb}{Numerotation of NA ref.raster cells}

\item{cell.Relpos.ngb}{Relative position of neighbours in the moving window}

\item{cell.Relpos.cross}{Numerotation of positions of neighbours of the moving window}
}
\value{
matrix with all neighbours of each cell that are not NA
}
\description{
Find neighbours without obstacles in different directions
}
