% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoR_dist_functions.R
\name{idw.dist}
\alias{idw.dist}
\title{Inverse distance calculation using custom distances}
\usage{
idw.dist(
  data,
  coords,
  locations,
  loc.dist,
  idp = 2,
  max.dist,
  nmin = 1,
  nmax,
  longlat = TRUE
)
}
\arguments{
\item{data}{vector of values to be interpolated at data coordinates}

\item{coords}{coordinates of observation points. Can also be an object of class
SpatialPoints*. Not necessary with loc.dist.}

\item{locations}{coordinates of points where to provide predictions. Can also be an object of class
SpatialPoints*. Not necessary with loc.dist.}

\item{loc.dist}{Matrix of distances between data (rows) and locations (cols)}

\item{idp}{specify the inverse distance weighting power. Default to 2.}

\item{max.dist}{Maximum distance radius for neighbours search in the idw interpolation.
maxdist is in km when longlat is TRUE (non-projected crs), in meters otherwise.
If loc.dist is specified, maxdist should be in same distance unit.}

\item{nmin}{if the number of nearest observations within distance maxdist is less than nmin, a missing value will be generated; see maxdist}

\item{nmax}{the number of nearest observations that should be used for a kriging prediction or simulation, where nearest is defined in terms of the space of the spatial locations. By default, all observations are used}

\item{longlat}{Logical Wether coordinates are not-projected (TRUE) or are projected (FALSE).}
}
\description{
Inverse distance calculation using custom distances
}
