% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoR_dist_functions.R
\name{idw.dist.special}
\alias{idw.dist.special}
\title{Special case idw for reduced time calculation}
\usage{
idw.dist.special(data, loc.dist.order, loc.dist.nmax, idp)
}
\arguments{
\item{data}{vector of data to interpolate on points. Refer to output of Prep_loc_dist}

\item{loc.dist.order}{matrix of nmax lines and nb points columns with position
of data to keep}

\item{loc.dist.nmax}{matrix of nmax lines and nb points columns with distances
between kept data and points position}

\item{idp}{specify the inverse distance weighting power. Default to 2.}
}
\description{
Special case idw for reduced time calculation
}
