% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoR_dist_functions.R
\name{krige.conv.dist}
\alias{krige.conv.dist}
\title{Kriging. Modified from function krige.conv in geoR library.}
\usage{
krige.conv.dist(geodata, coords = geodata$coords, data = geodata$data,
  locations, borders, krige, output, loc.dist)
}
\arguments{
\item{geodata}{a list containing elements \code{coords} and
    \code{data} as described next. Typically an object of the \code{\link{class}}
    \code{"geodata"} - a \pkg{geoR} data-set. If not provided the arguments
    \code{coords} and \code{data} must be provided instead.  }

\item{coords}{an \eqn{n \times 2}{n x 2} matrix or data-frame with the 2-D
    coordinates of the \eqn{n} data locations.
    By default it takes the
    component \code{coords} of the argument \code{geodata}, if provided.  }

\item{data}{a vector with \emph{n} data values. By default it takes the
    component \code{data} of the argument \code{geodata}, if provided.  }

\item{locations}{an \eqn{N \times 2}{N x 2} matrix or data-frame with the 2-D
    coordinates of the \eqn{N} prediction locations, or a list for which
    the first two components are used. Input is internally checked by the
    function \code{check.locations}.  }

\item{borders}{optional. By default reads the element \code{borders}
    from the \code{geodata} object, if present.
    Setting to \code{NULL} prevents this behavior.
    If a two column matrix defining a polygon is
    provided the prediction is performed only at locations inside this polygon. }

\item{krige}{a list defining the model components and the type of
    kriging. It can take an output to a call to \code{krige.control} or
    a list with elements as for the arguments in \code{krige.control}.
    Default values are assumed for arguments  or list elements not provided.
    See the description of arguments in `krige.control' below.  }

\item{output}{a list specifying output options.
    It can take an output to a call to \code{output.control} or
    a list with elements as for the arguments in \code{output.control}.
    Default values are assumed for arguments not provided.
    See documentation for
    \code{\link{output.control}} for further details.  }

\item{loc.dist}{is a n (data) x N (locations) matrix with distances between data points and prediction locations.}
}
\description{
Kriging. Modified from function krige.conv in geoR library.
}
