% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoR_dist_functions.R
\name{Prep_loc_dist}
\alias{Prep_loc_dist}
\title{Prepare loc.dist for idw special case}
\usage{
Prep_loc_dist(loc.dist, nmin = 1, nmax, max.dist)
}
\arguments{
\item{loc.dist}{Matrix of distances between data (rows) and locations (cols)}

\item{nmin}{if the number of nearest observations within distance maxdist is less than nmin, a missing value will be generated; see maxdist}

\item{nmax}{the number of nearest observations that should be used for a kriging prediction or simulation, where nearest is defined in terms of the space of the spatial locations. By default, all observations are used}

\item{max.dist}{Maximum distance radius for neighbours search in the idw interpolation.}
}
\description{
Prepare loc.dist for idw special case
}
