% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_distances_functions.R
\name{Pt2Pts.wo.obstacle}
\alias{Pt2Pts.wo.obstacle}
\title{Calculate distance from one point to a set of points using adjacency matrix}
\usage{
Pt2Pts.wo.obstacle(
  dref.from,
  dref.to,
  dref.r,
  r.ref,
  longlat,
  small.dist = TRUE,
  keep.path = FALSE
)
}
\arguments{
\item{dref.from}{object of class distref.data for a single point}

\item{dref.to}{object of class distref.data}

\item{dref.r}{object of class distref.raster corresponding to r.ref}

\item{r.ref}{reference raster (only used for its coordinates here)}

\item{longlat}{Logical. if FALSE, Euclidean distance, if TRUE Great Circle
distance. Defined from r.ref.}

\item{small.dist}{if TRUE, distances <= 1 paths are directly calculated.}

\item{keep.path}{Logical. Whether to keep SpatialLines of paths.
This allows real distances when 2 points are in the same raster cell or close.}
}
\value{
Vector of distances of length(B)
}
\description{
Calculate distance from one point to a set of points using adjacency matrix
}
\examples{
\dontrun{
# Need gdistance ??
}
}
