% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sph2car.R
\name{sph2car}
\alias{sph2car}
\title{Computes cartesian coordinates from long,lat geographical coordinates
Use of geoid with radius and flattening to calculate correct x,y,z coordinates}
\usage{
sph2car(long, lat, radius = 6378137, f = 1/298.257223563, deg = TRUE)
}
\arguments{
\item{long}{longitude values, can also contain a matrix of (long, lat), or
(long, lat and and radius) in that order.}

\item{lat}{latitude values.}

\item{radius}{major (equatorial) radius of the ellipsoid. The default value is for WGS84}

\item{f}{numeric. Ellipsoid flattening. The default value is for WGS84}

\item{deg}{Specifies if input is in degrees (default) or radians.}
}
\description{
Computes cartesian coordinates from long,lat geographical coordinates
Use of geoid with radius and flattening to calculate correct x,y,z coordinates
}
\details{
Use geosphere::refEllipsoids() for other radius and f possibilites.
}
