% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoR_dist_functions.R
\name{variog.dist}
\alias{variog.dist}
\title{Variogram calculation. Modified from geoR.}
\usage{
variog.dist(
  geodata,
  coords = geodata$coords,
  data = geodata$data,
  uvec = "default",
  breaks = "default",
  trend = "cte",
  lambda = 1,
  option = c("bin", "cloud", "smooth"),
  estimator.type = c("classical", "modulus"),
  nugget.tolerance,
  max.dist,
  dist.mat,
  pairs.min = 2,
  bin.cloud = FALSE,
  direction = "omnidirectional",
  tolerance = pi/8,
  unit.angle = c("radians", "degrees"),
  angles = FALSE,
  messages,
  ...
)
}
\arguments{
\item{geodata}{a list containing element \code{coords}
    as described next. Typically an object of the class
    \code{"geodata"} - a \pkg{geoR} data-set.
    If not provided the arguments
    \code{coords} must be provided instead.  }

\item{coords}{an \eqn{n \times 2}{n x 2} matrix containing
    coordinates of the \eqn{n} data locations in each row.
    Defaults to \code{geodata$coords}, if provided.}

\item{data}{a vector or matrix with data values.
    If a matrix is provided, each column is regarded as one variable or realization.
    Defaults to \code{geodata$data}, if provided.}

\item{uvec}{a vector with values used to define the variogram binning. Only
    used when \code{option = "bin"}.
    See \code{DETAILS} below for more details on how to specify the bins.
  }

\item{breaks}{a vector with values to define the variogram binning. Only
    used when \code{option = "bin"}.
    See \code{DETAILS} below for more details on how to specify the bins.
  }

\item{trend}{specifies the mean part of the model.
    See documentation of \code{\link[geoR]{trend.spatial}} for further details.
    Defaults to \code{"cte"}.  }

\item{lambda}{values of the Box-Cox transformation parameter.
    Defaults to \eqn{1} (no
    transformation). If another value is provided the variogram is
    computed after transforming the 
    data. A case of particular interest is \eqn{\lambda = 0}{lambda = 0}
    which corresponds to log-transformation.  }

\item{option}{defines the output type: the options \code{"bin"} returns values of
    binned variogram, \code{"cloud"} returns the variogram cloud and
    \code{"smooth"} returns the kernel smoothed variogram.
    Defaults to \code{"bin"}.}

\item{estimator.type}{\code{"classical"} computes the classical method of
    moments estimator.  \code{"modulus"} returns the variogram
    estimator suggested by Hawkins and Cressie (see Cressie, 1993, pg 75).
    Defaults to \code{"classical"}.  }

\item{nugget.tolerance}{a numeric value. Points which are separated
    by a distance less than this value
    are considered co-located. Defaults to zero.  }

\item{max.dist}{a numerical value defining the maximum distance for
    the variogram. Pairs of locations
    separated for distance larger than this value are ignored for the
    variogram calculation. If not provided defaults takes the maximum
    distance among all pairs of data locations.  }

\item{dist.mat}{Square matrix of distances between points of the dataset}

\item{pairs.min}{a integer number defining the minimum numbers of
    pairs for the bins.
    For \code{option = "bin"},
    bins with number of pairs smaller than this
    value are ignored. Defaults to \code{NULL}.  }

\item{bin.cloud}{logical. If \code{TRUE} and
    \code{option = "bin"} the cloud values for each class are
    included in the output. Defaults to \code{FALSE}.  }

\item{direction}{a numerical value for the directional (azimuth) angle. This
    used to specify directional variograms. Default defines the
    omnidirectional variogram. The value must be in the interval
    \eqn{[0, \pi]}{[0, pi]} radians (\eqn{[0, 180]} degrees).  }

\item{tolerance}{numerical value for the tolerance angle, when
    computing directional variograms. The value must be in the interval
    \eqn{[0, \pi/2]}{[0, pi/2]} radians (\eqn{[0, 90]} degrees).  Defaults to \eqn{\pi/8}{pi/8}.  }

\item{unit.angle}{defines the unit for the specification of angles in
    the two previous arguments. Options are \code{"radians"} and
    \code{"degrees"}, with default to \code{"radians"}. }

\item{angles}{Logical with default to \code{FALSE}. If \code{TRUE} the
    function also returns the angles between the pairs of points
    (unimplemented). }

\item{messages}{logical. Indicates whether
    status messages should be printed on the screen (or output device)
    while the function is running.  }

\item{...}{arguments to be passed to the function \code{\link{ksmooth}}, if
    \code{option = "smooth"}.  }
}
\description{
Variogram calculation. Modified from geoR.
}
