% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_distances_functions.R
\docType{class}
\name{distref.raster-class}
\alias{distref.raster-class}
\title{Class for distref.raster}
\description{
Class for distref.raster
}
\section{Slots}{

\describe{
\item{\code{r.ref.path}}{path of the raster}

\item{\code{r.ref.dim}}{dimensions of the raster}

\item{\code{r.ref.N}}{cells numerotation}

\item{\code{r.ref.NA.nb}}{numerotation of NA cells}

\item{\code{adj.ref.graph}}{adjacency graph}

\item{\code{path.w}}{weight (distance) of paths}
}}

