% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodist-package.r, R/geodist.R
\docType{package}
\name{geodist}
\alias{geodist}
\alias{geodist-package}
\alias{geodist}
\title{geodist.}
\usage{
geodist(x, y)
}
\arguments{
\item{x}{Rectangular object (matrix, \code{data.frame}, \pkg{tibble},
whatever) containing longitude and latitude coordinates.
#param y Optional second object which, if passed, results in distances
calculated between each object in \code{x} and each in \code{y}.}
}
\value{
If only \code{x} passed, a square matrix containing distances between
all items in \code{x}; otherwise if \code{y} is passed, the resultant matrix
has \code{nrow(x)} rows and \code{nrow(y)} columns.
}
\description{
geodist.

Convert one or two rectangular objects containing lon-lat coordinates into
matrix of geodesic distances.
}
