% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geodist_benchmark}
\alias{geodist_benchmark}
\title{geodist_benchmark}
\usage{
geodist_benchmark(lat = 0, d = 1, n = 100)
}
\arguments{
\item{lat}{Central latgitude where errors should be measured}

\item{d}{Distance in metres over which errors should be measured}

\item{n}{Number of random values used to generate estimates}
}
\description{
Benchmark errors for different geodist measures
}
\examples{
geodist_benchmark (0, 1, 100)
}
